添加后台框架，完成出题基本功能

+++ OnlineSurvey/QuickWeb/Controllers/Common/BaseController.cs
@@ -787,7 +787,7 @@ namespace QuickWeb.Controllers.Common
         #endregion
 
         #region 通用Id非法校验
-        protected bool IsIllegalId(int? id)
+        protected bool IsIllegalId(long? id)
         {
             if (id == null || id <= 0)
                 return true;
